move into localize directory
