+++ README.md
@@ -71,7 +71,7 @@ npm install @brightspace-ui/core
   * [FormElementMixin](components/form/docs/form-element-mixin.md): allow components to participate in forms and validation
   * [InteractiveMixin](mixins/interactive-mixin.md): enables toggling interactive elements inside of nested grids
   * [LabelledMixin](mixins/labelled-mixin.md): label custom elements by referencing elements across DOM scopes
-  * [LocalizeMixin](mixins/localize-mixin.md): localize text in your components
+  * [LocalizeMixin](mixins/localize/): localize text in your components
   * [ProviderMixin](mixins/provider-mixin.md): provide and consume data across elements in a DI-like fashion
   * [RtlMixin](mixins/rtl-mixin.md): enable components to define RTL styles
   * [SkeletonMixin](components/skeleton/): make components skeleton-aware
