+++ index.html
@@ -180,7 +180,7 @@
 	<ul>
 		<li><a href="mixins/async-container/demo/async-container.html">async-container-mixin</a></li>
 		<li><a href="mixins/demo/labelled-mixin.html">labelled-mixin</a></li>
-		<li><a href="mixins/demo/localize-mixin.html">localize-mixin</a></li>
+		<li><a href="mixins/localize/demo/localize-mixin.html">localize-mixin</a></li>
 		<li><a href="components/skeleton/demo/skeleton-mixin.html">skeleton-mixin</a></li>
 	</ul>
